move short channel id to basetypes dir
