+++ basetypes/short_channel_id.go
@@ -1,4 +1,4 @@
-package main
+package basetypes
 
 import (
 	"fmt"
